splitting up zanj
