+++ muutils/zanj/zanj.py
@@ -40,222 +40,8 @@ ZANJitem = Union[
 
 ExternalsLoadingMode = Literal["lazy", "full"]
 
-
-
-def jsonl_metadata(data: list[JSONitem]) -> dict:
-	return {
-		"data[0]": data[0],
-		"len(data)": len(data),
-	}
-
-
-def zanj_external_serialize(
-		jser: "ZANJ", 
-		data: Any, 
-		path: ObjectPath,
-		item_type: ExternalItemType,
-	) -> JSONitem:
-	"""stores a numpy array or jsonl externally in a ZANJ object
-	
-	# Parameters:
-	 - `jser : ZANJ`
-	 - `arr : NDArray`
-	
-	# Returns:
-	 - `JSONitem` 
-	   json data with reference
-
-	# Modifies:
-	 - modifies `jser._externals`
-	"""	
-	# get the path, make sure its unique
-	joined_path: str = "/".join([str(p) for p in path])
-	if joined_path in jser._externals:
-		raise ValueError(f"external path {joined_path} already exists!")
-
-	# process the data if needed, assemble metadata
-	data_new: Any = data
-	output: dict = {
-		"__format__": f"external:{EXTERNAL_ITEMS_EXTENSIONS[item_type]}",
-		"$ref": joined_path,
-	}
-	if item_type == "ndarray":
-		# check type
-		data_type_str: str = str(type(data))
-		if data_type_str == "<class 'torch.Tensor'>":
-			# detach and convert
-			data_new = data.detach().cpu().numpy()		
-		elif data_type_str == "<class 'numpy.ndarray'>":
-			pass
-		else:
-			# if not a numpy array, except
-			raise TypeError(f"expected numpy.ndarray, got {data_type_str}")
-		# get metadata
-		output.update(arr_metadata(data))
-	elif item_type.startswith("jsonl"):
-		if isinstance(data, pd.DataFrame):
-			output["columns"] = data.columns.tolist()
-			data_new = data.to_dict(orient="records")
-		elif isinstance(data, (list, tuple, Iterable)):
-			data_new = [
-				jser.json_serialize(item, path + (i,))
-				for i, item in enumerate(data)
-			]
-		else:
-			raise TypeError(f"expected list or pd.DataFrame, got {type(data)}")
-			
-		output.update(jsonl_metadata(data_new))
-			
-	# store the item for external serialization
-	jser._externals[joined_path] = ExternalItem(
-		item_type=item_type,
-		data=data_new,
-		path=path,
-	)
-
-	return output
-
-def zanj_serialize_torchmodule(
-		jser: "ZANJ",
-		data: "torch.nn.Module",
-		path: ObjectPath,
-	) -> JSONitem:
-	"""serialize a torch module to zanj
-	
-	we want to save:
-	- class name, docstring, __mro__
-	- code of `.forward()` method
-	- code of `.__init__()` method
-	- state dict, in accordance with zanj parameters for storing arrays
-	- `_modules` with modules as strings
-	- __dict__ with values printed (not fully serialized)
-	"""
-
-	output: dict = {
-		"__format__" : "torchmodule",
-		"__class__" : str(data.__class__.__name__),
-		"__doc__" : string_as_lines(data.__doc__),
-		"__mro__" : [str(c.__name__) for c in data.__class__.__mro__],
-		"__init__" : string_as_lines(inspect.getsource(data.__init__)),
-		"forward" : string_as_lines(inspect.getsource(data.forward)),
-		"state_dict" : jser.json_serialize(data.state_dict(), path + ("state_dict",)),
-		"_modules" : {k: str(v) for k, v in data._modules.items()},
-		"__dict__" : {k: str(v) for k, v in data.__dict__.items()},
-	}
-
-	return jser.json_serialize(output)
-
-
 ZANJ_MAIN: str = "__zanj__.json"
 ZANJ_META: str = "__zanj_meta__.json"
-
-
-DEFAULT_SERIALIZER_HANDLERS_ZANJ: MonoTuple[SerializerHandler] = (
-	(
-		SerializerHandler(
-			check = lambda self, obj, path: (
-				isinstance(obj, np.ndarray) 
-				and obj.size >= self.external_array_threshold
-			),
-			serialize = lambda self, obj, path: zanj_external_serialize(self, obj, path, item_type="ndarray"),
-			desc = "external:ndarray",
-		),
-		SerializerHandler(
-			check = lambda self, obj, path: (
-				str(type(obj)) == "<class 'torch.Tensor'>" 
-				and int(obj.nelement()) >= self.external_array_threshold
-			),
-			serialize = lambda self, obj, path: zanj_external_serialize(self, obj, path, item_type="ndarray"),
-			desc = "external:ndarray:torchtensor",
-		),
-		SerializerHandler(
-			check = lambda self, obj, path: isinstance(obj, (list, tuple, pd.DataFrame)) and len(obj) >= self.external_table_threshold,
-			serialize = lambda self, obj, path: zanj_external_serialize(self, obj, path, item_type="jsonl"),
-			desc = "external:jsonl",
-		),
-		SerializerHandler(
-			check = lambda self, obj, path: "<class 'torch.nn.modules.module.Module'>" in [str(t) for t in obj.__class__.__mro__],
-			serialize = lambda self, obj, path: zanj_serialize_torchmodule(self, obj, path),
-			desc = "torch.nn.Module",
-		),
-	)
-	+ DEFAULT_HANDLERS
-)
-
-
-@dataclass
-class LoaderHandler:
-	"""handler for loading an object from a json file"""
-	# (json_data, path) -> whether to use this handler
-	check: Callable[[JSONitem, ObjectPath], bool]
-	# function to load the object
-	load: Callable[[JSONitem, ObjectPath], Any]
-	# description of the handler
-	desc: str = "(no description)"
-
-@dataclass
-class ZANJLoaderHandler:
-	"""handler for loading an object from a ZANJ archive (takes ZANJ object as first arg)"""
-	# (zanj_obi, json_data, path) -> whether to use this handler
-	check: Callable[["ZANJ", JSONitem, ObjectPath], bool]
-	# function to load the object (zanj_obj, json_data, path) -> loaded_obj
-	load: Callable[["ZANJ", JSONitem, ObjectPath], Any]
-	# TODO: add name/unique identifier, `desc` should be human-readable and more detailed
-	# description of the handler
-	desc: str = "(no description)"
-
-	@classmethod
-	def from_LoaderHandler(cls, lh: LoaderHandler):
-		"""wrap a standard LoaderHandler to make it compatible with ZANJ"""
-		return cls(
-			check = lambda zanj, json_item, path: lh.check(json_item, path),
-			load = lambda zanj, json_item, path: lh.load(json_item, path),
-			desc = lh.desc,
-		)
-
-DEFAULT_LOADER_HANDLERS: tuple[LoaderHandler] = (
-	LoaderHandler(
-		check = lambda json_item, path: (
-			isinstance(json_item, dict)
-			and "__format__" in json_item
-			and json_item["__format__"] == "array_list_meta"
-		),
-		load = lambda json_item, path: (
-			np.array(json_item["data"], dtype=json_item["dtype"]).reshape(json_item["shape"])
-		),
-		desc = "array_list_meta loader",
-	),
-	LoaderHandler(
-		check = lambda json_item, path: (
-			isinstance(json_item, dict)
-			and "__format__" in json_item
-			and json_item["__format__"] == "array_hex_meta"
-		),
-		load = lambda json_item, path: (
-			np.frombuffer(bytes.fromhex(json_item["data"]), dtype=json_item["dtype"]).reshape(json_item["shape"])
-		),
-		desc = "array_hex_meta loader",
-	),
-)
-
-
-DEFAULT_LOADER_HANDLERS_ZANJ: tuple[ZANJLoaderHandler] = (
-	ZANJLoaderHandler(
-		check = lambda zanj, json_item, path: (
-			isinstance(json_item, dict)
-			and "__format__" in json_item
-			and json_item["__format__"].startswith("external:")
-		),
-		load = lambda zanj, json_item, path: (
-			zanj._externals[json_item["key"]]
-		)
-	),
-) + tuple(
-	ZANJLoaderHandler.from_LoaderHandler(lh) 
-	for lh in DEFAULT_LOADER_HANDLERS
-)
-
-
 
 class ZANJ(JsonSerializer):
 	"""Zip up: Arrays in Numpy, JSON for everything else
@@ -382,115 +168,4 @@ class ZANJ(JsonSerializer):
 
 	def read(self, path: Union[str, Path], externals_mode: ExternalsLoadingMode = "lazy") -> Any:
 		"""load the object from a ZANJ archive"""
-		
-
-
-
-@dataclass
-class ZANJLoaderTreeNode:
-	"""acts like a regular dictionary or list, but applies loaders to items
-	
-	does this by passing `_parent` down the stack, whenever you get an item from the container.
-	"""
-	_parent: "LoadedZANJ"
-	_data: dict[str, JSONitem]|list[JSONitem]
-
-	def __getitem__(self, key: str|int) -> Any:
-		# get value, check key types
-		val: JSONitem
-		if (
-				(isinstance(self._data, list) and isinstance(key, int))
-				or (isinstance(self._data, dict) and isinstance(key, str))
-			):
-			val = self._data[key]
-		else:
-			raise KeyError(f"invalid key type {type(key) = }, {key = } for {self._data = } with {type(self._data) = }")
-
-		# get value or dereference
-		if key == "$ref":
-			val = self._parent._externals[val]
-		else:
-			val = self._data[key]
-
-		if isinstance(val, (dict,list)):
-			return ZANJLoaderTreeNode(_parent=self._parent, _data=val)
-		else:
-			return val
-
-
-class LazyExternalLoader:
-	"""lazy load np arrays or jsonl files, similar tp NpzFile from np.lib
-	
-	initialize with zipfile object and zanj metadata (for list of externals)
-	"""
-
-	def __init__(
-			self,
-			zipf: zipfile.ZipFile,
-			zanj_meta: JSONitem,
-		):
-		self._zipf: zipfile.ZipFile = zipf
-		self._zanj_meta: JSONitem = zanj_meta
-		# (path, item_type) pairs
-		self._externals_types: dict[str, str] = {
-			key : val.item_type
-			for key, val in zanj_meta["externals_info"].items()
-		}
-
-		# validate by checking each external file exists
-		for key, item_type in self._externals_types.items():
-			fname: str = f"{key}.{EXTERNAL_ITEMS_EXTENSIONS[item_type]}"
-			if fname not in zipf.namelist():
-				raise ValueError(f"external file {fname} not found in archive")
-
-
-	def __getitem__(self, key: str) -> Any:
-		if key in self._externals_types:
-			path, item_type = key
-			with self._zipf.open(path, "rb") as fp:
-				return EXTERNAL_LOAD_FUNCS[item_type](fp)
-
-
-
-class LoadedZANJ:
-	"""object loaded from ZANJ archive. acts like a dict."""
-	def __init__(
-		self,
-		# config
-		path: str|Path,
-		zanj: ZANJ,
-		loader_handlers: MonoTuple[ZANJLoaderHandler],
-		externals_mode: ExternalsLoadingMode = "lazy",
-	) -> None:
-
-		self._path: str = str(path)
-		self._zanj: ZANJ = zanj
-		self._externals_mode: ExternalsLoadingMode = externals_mode
-
-		self._zipf: zipfile.ZipFile = zipfile.ZipFile(file=self._path, mode="r")
-
-		self._meta: JSONitem = json.load(self._zipf.open(ZANJ_META, "rt"))
-		self._json_data: ZANJLoaderTreeNode = ZANJLoaderTreeNode(
-			_parent = self,
-			_data = json.load(self._zipf.open(ZANJ_MAIN, "rt"))
-		)
-
-		self._externals: LazyExternalLoader|dict
-		if externals_mode == "lazy":
-			self._externals = LazyExternalLoader(
-				zipf=self._zipf,
-				zanj_meta=self._meta,
-			)
-		elif externals_mode == "full":
-			self._externals = {
-				key : self._load_external(key, val)
-				for key, val in self._meta["externals_info"].items()
-			}
-
-		
-	
-	def __getitem__(self, key: str) -> Any:
-		"""get the value of the given key"""
-
-
-
+		pass
